The 'covidcast' package.

A DESCRIPTION OF THE PACKAGE

covidcast returns a base configuration of the model with the default set of priors, and no input data. This configuration, after adding input data (see input_cases, input_deaths, and input_fracpos, represents a valid model configuration that can be passed to run.

covidcast(
  N_days,
  N_days_before = 28,
  chains = 3,
  iter = 1500,
  thin = 1,
  rho = 1,
  seed = 42
)

Arguments

N_days
A positive integer. The number of days of input data being modeled. This should always be set to the number of days in your input data.

N_days_before
A positive integer. How many days before the first day of model data should be modeled?

chains
The number of chains to use during MCMC, as passed to sampling.

iter
The number of iterations to run during MCMC, as passed to sampling.

thin
A positive integer to specify period for saving samples, as passed to sampling.

rho
A number in (0, 1]. Needs documentation.

seed
A number. The random number generator seed for use in sampling.

Value

An S3 object of type covidcast. This can be passed to run to execute the model. This object can also be saved to disk using saveRDS to enable reproducibility across platforms or sessions.

References

Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org

Examples

covidcast(N_days = 50, seed = 42)

#> Priors:
#> 
#> 	log_new_inf_0_mu	0
#> 	log_new_inf_0_sd	10
#> 	deriv1_log_new_inf_sd	0.5
#> 	deriv2_log_new_inf_sd	0.05
#> 	p_sym_if_inf	[alpha]	5.9
#> 	p_sym_if_inf	[beta]	4.1
#> 	p_sev_if_sym	[alpha]	3.1
#> 	p_sev_if_sym	[beta]	6.9
#> 	p_die_if_sev	[alpha]	0.3
#> 	p_die_if_sev	[beta]	9.7
#> 	p_die_if_sym	[alpha]	1.12
#> 	p_die_if_sym	[beta]	85.1
#> 	inf_prg_delay	[shape]	5.202
#> 	inf_prg_delay	[rate]	0.946
#> 	sym_prg_delay	[shape]	5.147
#> 	sym_prg_delay	[rate]	0.468
#> 	sev_prg_delay	[shape]	2.383
#> 	sev_prg_delay	[rate]	0.27
#> 	cas_rep_delay_shp	[alpha]	2.2
#> 	cas_rep_delay_shp	[beta]	1
#> 	die_rep_delay_shp	[alpha]	2.2
#> 	die_rep_delay_shp	[beta]	1
#> 	p_diag_if_sym	[alpha]	2
#> 	p_diag_if_sym	[beta]	2
#> 	p_diag_if_sev	[alpha]	1.5
#> 	p_diag_if_sev	[beta]	1
#> 	weekend_eff	[alpha]	1.5
#> 	weekend_eff	[beta]	1.5
#> 
#> Inputs:
#> 
#> 	[ ❌ ]	Cases
#> 	[ ❌ ]	Deaths
#> 	[ ❌ ]	Fraction positive