Obtain COVIDcast metadata

Obtains a data frame of metadata describing all publicly available data streams from the COVIDcast API.

covidcast_meta()

Value

Data frame containing one row per signal, with the following columns:

data_source

Data source name.

signal

Signal name.

min_time

First day for which this signal is available.

max_time

Most recent day for which this signal is available.

geo_type

Geographic level for which this signal is available, such as county, state, msa, or hrr. Most signals are available at multiple geographic levels and will hence be listed in multiple rows with their own metadata.

time_type

Temporal resolution at which this signal is reported. "day", for example, means the signal is reported daily.

num_locations

Number of distinct geographic locations available for this signal. For example, if geo_type is county, the number of counties for which this signal has ever been reported.

min_value

Smallest value that has ever been reported.

max_value

Largest value that has ever been reported.

mean_value

Arithmetic mean of all reported values.

stdev_value

Sample standard deviation of all reported values.

max_issue

Most recent issue date for this signal.

min_lag

Smallest lag from observation to issue, in time_type units

max_lag

Largest lag from observation to issue, in time_type units

References

COVIDcast API sources and signals documentation: https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html

See also