Obtain multiple COVIDcast signals in one data frame

This convenience function uses covidcast_signal() to obtain multiple signals, potentially from multiple data sources.

covidcast_signals(
  data_source,
  signal,
  start_day = NULL,
  end_day = NULL,
  geo_type = c("county", "hrr", "msa", "dma", "state"),
  geo_values = "*",
  as_of = NULL,
  issues = NULL,
  lag = NULL
)

Details

The argument structure is just as in covidcast_signal(), except the first four arguments data_source, signal, start_day, end_day are permitted to be vectors. The returned object is a list of covidcast_signal data frames, of length N = max(length(data_source), length(signal)). The first two arguments data_source, signal are recycled appropriately, in the calls to covidcast_signal(); see example below. The next two arguments start_day, end_day, unless NULL, must be either length 1 or N.

See also

Examples

if (FALSE) {
## Fetch USAFacts confirmed cases and deaths over the same time period
covidcast_signals("usa-facts", signal=c("confirmed_incidence_num",
                                        "deaths_incidence_num"),
                   start_day = "2020-08-15", end_day = "2020-10-01")
}