Get county or metropolitan area names from FIPS or CBSA codes

Look up county or metropolitan area names by FIPS or CBSA codes, respectively; these functions are based on grep(), and hence allow for regular expressions.

fips_to_name(
  code,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  ties_method = c("first", "all")
)

cbsa_to_name(
  code,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  ties_method = c("first", "all")
)

Arguments

code
Vector of FIPS or CBSA codes to look up.

ignore.case, perl, fixed
Arguments to pass to grep(), with the same defaults as in the latter function. Hence, by default, regular expressions are used; to match against a fixed string (no regular expressions), set fixed = TRUE.

ties_method
If "first", then only the first match for each code is returned. If "all", then all matches for each code are returned.

Value

A vector of county or metro names if ties_method equals "first", and a list of county or names otherwise.

See also

Examples

fips_to_name("42003")

#>              42003 
#> "Allegheny County" 

cbsa_to_name("38300")

#>            38300 
#> "Pittsburgh, PA" 

fips_to_name("4200", ties_method = "all")

#> [[1]]
#>              42000              42001              42003              42005 
#>     "Pennsylvania"     "Adams County" "Allegheny County" "Armstrong County" 
#>              42007              42009 
#>    "Beaver County"   "Bedford County" 
#> 

# Count the number of counties, grouped by first two digits of FIPS code
# (which identify states):
unlist(lapply(fips_to_name(sprintf("%02d", 1:99), ties = "all"), length))

#>  [1] 462 252 286 203 337 228 252 220 245 405 325 207 335  92 178 135 251 154 216
#> [20] 311 269  93 107  36  98  95 171  95 193 191 219  73  93  27 103  69 168  59
#> [39] 154 207 107  73  65  11 112  72 158 255  92 229 238   1  97  57 128  24  57
#> [58]   2  57 276 126   1  55   2  54   1  54   4  53 210 231   3  53   1  53   1
#> [77]  51   2  49 260 187  52  99  51  53   2  46   1  44 197 178  15  42   1  46
#> [96]   1  43   1  42