Metro area population data

Data set on metropolitan area populations, from the 2019 US Census. This includes metropolitan and micropolitan statistical areas, although the COVIDcast API only supports fetching data from metropolitan statistical areas.

msa_census

Format

A data frame with 2797 rows, each representing one core-based statistical area (including metropolitan and micropolitan statistical areas, county or county equivalents, and metropolitan divisions). There are many columns. The most crucial are:

CBSA

Core Based Statistical Area code. These are unique identifiers used, for example, as the geo_values argument to covidcast_signal() when requesting data from specific metro areas (with geo_type = 'msa').

MDIV

Metropolitan Division code

STCOU

State and county code

NAME

Name or title of the area.

LSAD

Legal/Statistical Area Description, identifying if this is a metropolitan or micropolitan area, a metropolitan division, or a county.

STATE

State FIPS code.

POPESTIMATE2019

Estimate of the area's resident population as of July 1, 2019.

Source

United States Census Bureau, at https://www2.census.gov/programs-surveys/popest/datasets/2010-2019/metro/totals/cbsa-est2019-alldata.csv

References

Census Bureau documentation of all columns and their meaning: https://www2.census.gov/programs-surveys/popest/datasets/2010-2019/metro/totals/cbsa-est2019-alldata.pdf

See also