Obtain multiple signals in one data frame.

This convenience function uses covidcast_signal() to obtain multiple signals, potentially from multiple data sources, in one data frame. See the covidcast_signal() documentation for further details.

covidcast_signals(
  signals,
  start_day = NULL,
  end_day = NULL,
  geo_type = c("county", "hrr", "msa", "dma", "state"),
  geo_values = "*",
  as_of = NULL,
  issues = NULL,
  lag = NULL
)

Arguments

signals
Data frame with two columns: data_source and signal. Each row specifies one signal to be fetched from the COVIDcast API.

start_day
Query data beginning on this date. Date object, or string in the form "YYYY-MM-DD". If start_day is NULL, defaults to first day data is available for this signal.

end_day
Query data up to this date, inclusive. Date object or string in the form "YYYY-MM-DD". If end_day is NULL, defaults to the most recent day data is available for this signal.

geo_type
The geography type for which to request this data, such as "county" or "state". Defaults to "county". See the geographic coding documentation for details on which types are available.

geo_values
Which geographies to return. The default, "*", fetches all geographies. To fetch specific geographies, specify their IDs as a vector or list of strings. See the geographic coding documentation for details on how to specify these IDs.

as_of
Fetch only data that was available on or before this date, provided as a Date object or string in the form "YYYY-MM-DD". If NULL, the default, return the most recent available data.

issues
Fetch only data that was published or updated ("issued") on these dates. Provided as either a single Date object (or string in the form "YYYY-MM-DD"), indicating a single date to fetch data issued on, or a vector specifying two dates, start and end. In this case, return all data issued in this range. There may be multiple rows for each observation, indicating several updates to its value. If NULL, the default, return the most recently issued data.

lag
Integer. If, for example, lag=3, fetch only data that was published or updated exactly 3 days after the date. For example, a row with time_value of June 3 will only be included in the results if its data was issued or updated on June 6. If NULL, the default, return the most recently issued data regardless of its lag.

Value

Data frame with matching data. Each row is one observation of one signal on one day in one geographic location. Contains the following columns:

data_source

Data source from which this observation was obtained.

signal

The signal from which this observation was obtained.

geo_value

identifies the location, such as a state name or county FIPS code

time_value

a Date object identifying the date of this observation

issue

a Date object identifying the date this estimate was issued. For example, an estimate with a time_value of June 3 might have been issued on June 5, after the data for June 3rd was collected and ingested into the API.

lag

an integer giving the difference between issue and time_value, in days.

value

the signal quantity requested. For example, in a query for the confirmed_cumulative_num signal from the usa-facts source, this would be the cumulative number of confirmed cases in the area, as of the time_value.

stderr

the value's standard error, if available

sample_size

indicates the sample size available in that geography on that day; sample size may not be available for all signals, due to privacy or other constraints, in which case they will be NA.

direction

uses a local linear fit to estimate whether the signal in this region is currently increasing or decreasing (reported as -1 for decreasing, 1 for increasing, and 0 for neither).

Consult the signal documentation for more details on how values and standard errors are calculated for specific signals.
Details

Deliberately not exported, because.

References

COVIDcast API documentation: https://cmu-delphi.github.io/delphi-epidata/api/covidcast.html

Documentation of all COVIDcast sources and signals: https://cmu-delphi.github.io/delphi-epidata/api/covidcast_signals.html

COVIDcast public map: https://covidcast.cmu.edu/

See also

Examples

if (FALSE) {
signals <= data.frame(data_source=c("jhu-csse", "fb-survey"),
                      signal=c("confirmed_incidence_num", "smoothed_cli"))
covidcast_signals(signals, "2020-07-01", "2020-07-14", geo_type="state")
}